# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Recipe Collection demo using Express and Swagger
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/v1/recipe/{id}:
    # our controller name
    x-swagger-router-controller: recipe
    get:
      operationId: getOne
      description: get a recipe
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRecipeResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update a recipe
      # define the parameters
      parameters:
        - name: id
          description: Recipe id
          type: string
          in: path
          required: true
        - name: title
          description: recipe title
          in: body
          required: true
          schema:
            $ref: "#/definitions/Recipe"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delRecipe
      description: delete a recipe
      # define the parameters
      parameters:
        - name: id
          description: Recipe id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GetRecipeResponse:
    required:
      - id
      - title
      - ingredients
      - method
    properties:
      id:
        type: string
      title: 
        type: string
      ingredients:
        type: array
        items:
          type: string
      method:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Recipe:
    type: object
    properties:
      title:
        type: string
        description: recipe title
      cuisine:
        type: string
        description: recipe cuisine type
      ingredients:
        type: array
        items:
          type: string
      method:
        type: string
      difficulty:
        type: string
      cookingTime:
        type: integer
    required:
      - title
      - ingredients
      - method